+++ TxDOTAgent/TxDOTLegacyCadWriter.cpp
@@ -317,7 +317,7 @@ int TxDOT_WriteCADDataForGirder(FILE *fp, IBroker* pBroker, const CGirderKey& gi
 	   //----- COL 0a ---- 
 	   workerB.WriteFloat64(roadwayWidth,_T("RoadW"),7,5,_T("%5.2f"));
 	   //----- COL 0b ----- 
-	   workerB.WriteInt16((Int16)nGirders,_T("Ng "),5,3,_T("%3d"));
+	   workerB.WriteInt32((Int32)nGirders,_T("Ng "),5,3,_T("%3d"));
 	   //----- COL 0c ----- 
 	   workerB.WriteFloat64(girderSpacing,_T("Spcng"),7,5,_T("%5.2f"));
    }
@@ -477,11 +477,11 @@ int TxDOT_WriteCADDataForGirder(FILE *fp, IBroker* pBroker, const CGirderKey& gi
 	//----- COL 4 ----- 
    workerB.WriteString(strandPat,_T("N"),6,1,_T("%1s"));
 	//----- COL 5 ----- 
-   workerB.WriteInt16((Int16)strandNum,_T("Ns"),6,3,_T("%3d"));
+   workerB.WriteInt32((Int32)strandNum,_T("Ns"),6,3,_T("%3d"));
 	//----- COL 6 ----- 
    workerB.WriteStringEx(strandSize,_T("Size"),0,4,1,_T("%4s"));
 	//----- COL 7 ----- 
-   workerB.WriteInt16(strandStrength,_T("Strn"),5,3,_T("%3d"));
+   workerB.WriteInt32(strandStrength,_T("Strn"),5,3,_T("%3d"));
 	//----- COL 8 ----- 
    workerB.WriteFloat64(strandEccCL,_T("EccCL"),7,5,_T("%5.2f"));
 	//----- COL 9 ----- 
@@ -562,7 +562,7 @@ int TxDOT_WriteCADDataForGirder(FILE *fp, IBroker* pBroker, const CGirderKey& gi
 
       // output
       //----- COL 10 ---- 
-      workerB.WriteInt16((Int16)dstrandNum,_T("Nh"),(isExtendedVersion? 4:5),2,_T("%2d"));
+      workerB.WriteInt32((Int32)dstrandNum,_T("Nh"),(isExtendedVersion? 4:5),2,_T("%2d"));
 	   //----- COL 11 ---- 
       workerB.WriteFloat64(dstrandToEnd,_T("ToEnd"),5,4,_T("%4.1f"));
       workerB.WriteFloat64(dstrandToCL,_T("ToCL"),5,4,_T("%4.1f"));
@@ -595,7 +595,7 @@ int TxDOT_WriteCADDataForGirder(FILE *fp, IBroker* pBroker, const CGirderKey& gi
          dstrandToCL = ::ConvertFromSysUnits( value+Hg, unitMeasure::Inch );
 
          // output
-         workerB.WriteInt16((Int16)numRaisedStraightStrands,_T("Nh"),(isExtendedVersion? 4:5),2,_T("%2d"));
+         workerB.WriteInt32((Int32)numRaisedStraightStrands,_T("Nh"),(isExtendedVersion? 4:5),2,_T("%2d"));
          workerB.WriteFloat64(dstrandToEnd,_T("ToEnd"),5,4,_T("%4.1f"));
          workerB.WriteFloat64(dstrandToCL,_T("ToCL"),5,4,_T("%4.1f"));
       }
@@ -618,7 +618,7 @@ int TxDOT_WriteCADDataForGirder(FILE *fp, IBroker* pBroker, const CGirderKey& gi
    workerB.WriteFloat64(designLoadTensileStress,_T(" ftens "),10,6,_T("%6.3f"));
 
 	//----- COL 16 ---- 
-   workerB.WriteInt16(reqMinUltimateMomentCapacity,_T("ultMo"),9,5,_T("%5d"));
+   workerB.WriteInt32(reqMinUltimateMomentCapacity,_T("ultMo"),9,5,_T("%5d"));
 	//----- COL 17 ---- 
    workerB.WriteFloat64(momentDistFactor,_T("LLDFm"),7,5,_T("%5.3f"));
 	//----- COL 17aa ---- 
@@ -748,21 +748,21 @@ void TxDOTCadWriter::WriteInitialData(CadWriterWorkerBee& workerB)
    if (m_NumDebonded > 0)
    {
       // write out debonding data for bottom row
-      workerB.WriteInt16((Int16)m_NumDebonded,_T("Ndb"),4,2,_T("%2d"));
+      workerB.WriteInt32((Int32)m_NumDebonded,_T("Ndb"),4,2,_T("%2d"));
 
       if (m_Rows.empty() || m_OutCome==SectionMismatch || m_OutCome==TooManySections || m_OutCome==SectionsNotSymmetrical)
       {
          // row height, srands in row, and debonds in row are zero
 	      workerB.WriteFloat64(0.0,_T("Debnd"),7,5,_T("%5.2f"));
-         workerB.WriteInt16(0,_T("   "),6,2,_T("%2d"));
-         workerB.WriteInt16(0,_T("   "),6,2,_T("%2d"));
+         workerB.WriteInt32(0,_T("   "),6,2,_T("%2d"));
+         workerB.WriteInt32(0,_T("   "),6,2,_T("%2d"));
 
          if (m_Rows.empty())
          {
             // no use searching for nothing
             for (int i=0; i<5; i++)
             {
-               workerB.WriteInt16(0,_T("  "),4,2,_T("%2d"));
+               workerB.WriteInt32(0,_T("  "),4,2,_T("%2d"));
             }
          }
          else
@@ -889,20 +889,20 @@ void TxDOTCadWriter::WriteRowData(CadWriterWorkerBee& workerB, const RowData& ro
    }
 
    // total strands in row
-   workerB.WriteInt16((Int16)row.m_NumTotalStrands,_T("Nsr"),6,2,_T("%2d"));
+   workerB.WriteInt32((Int32)row.m_NumTotalStrands,_T("Nsr"),6,2,_T("%2d"));
 
    // num debonded strands in row
-   Int16 nsr = CountDebondsInRow(row);
-   workerB.WriteInt16(nsr,_T("Ndb"),6,2,_T("%2d"));
+   Int32 nsr = CountDebondsInRow(row);
+   workerB.WriteInt32(nsr,_T("Ndb"),6,2,_T("%2d"));
 
 	//----- COL 14-23 ---- 
    // we have 5 columns to write no matter what
    SectionListConstIter scit = row.m_Sections.begin();
 
    TCHAR buff[4];
-   for (Int16 icol=0; icol<5; icol++)
+   for (Int32 icol=0; icol<5; icol++)
    {
-      Int16 db_cnt = 0;
+      Int32 db_cnt = 0;
 
       if (scit!= row.m_Sections.end())
       {
@@ -918,7 +918,7 @@ void TxDOTCadWriter::WriteRowData(CadWriterWorkerBee& workerB, const RowData& ro
 
       _stprintf_s(buff,sizeof(buff)/sizeof(TCHAR),_T("%2d"),icol+1);
 
-      workerB.WriteInt16(db_cnt,buff,4,2,_T("%2d"));
+      workerB.WriteInt32(db_cnt,buff,4,2,_T("%2d"));
    }
 
    ATLASSERT(scit==row.m_Sections.end()); // we didn't find all of our sections - bug
